+++ tests/manual/_template0_camera-environment-map.tsx
@@ -2,23 +2,21 @@
 /* eslint-disable global-require */
 /* eslint-disable import/no-unresolved */
 // eslint-disable-next-line import/no-extraneous-dependencies
-import * as THREE from "three";
-import React, { useState } from "react";
+import React from "react";
 import { render } from "react-dom";
 import { ZapparCamera, ZapparCanvas } from "../../src/index";
 
-const App = () => {
-  const [envMap, setEnvMap] = useState<THREE.Texture>();
-
+function App() {
   return (
     <ZapparCanvas>
-      <ZapparCamera useEnvironmentMap={setEnvMap} />
+      <ZapparCamera environmentMap />
+      {/* <CameraEnvironmentMap attach="environment" /> */}
       <mesh position={[0, 0, -5]}>
         <sphereBufferGeometry />
-        <meshStandardMaterial metalness={1} roughness={0} envMap={envMap} />
+        <meshStandardMaterial metalness={1} roughness={0} />
       </mesh>
       <directionalLight position={[2.5, 8, 5]} intensity={1.5} />
     </ZapparCanvas>
   );
-};
+}
 render(<App />, document.getElementById("root"));